refactor: modularize client code into separate pool module
Move ConnectionPoolState, ErrorStats, and ResponseResult to dedicated
src/client/pool.rs module. Re-export from mod.rs for clean public API.
Improves code organization and separation of concerns.
