+++ src/client/mod.rs
@@ -1,1 +1,5 @@
 pub mod h3_client;
+pub mod pool;
+
+// Re-export public types
+pub use pool::{ConnectionPoolState, ErrorStats, ResponseResult};
